==> 01_solutions.Rmd <==
---
title: "session 1 exercise solutions"
author: "Tim Riffe"
date: "2026-07-07"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
# 1. Merge `B` and `Ex`, matching ages, but taking care that births are matched to female exposures.

## Load the data

I got these urls from the _raw_ button of the respective files on github:

```{r, message = FALSE}
library(tidyverse)
library(janitor)
B <- read_csv("https://github.com/timriffe/BSSD2025Module2/raw/master/data/ES_B.csv.gz",
              show_col_types = FALSE)
P <- read_csv("https://github.com/timriffe/BSSD2025Module2/raw/master/data/ES_P.csv.gz",
              show_col_types = FALSE)
```
## Calculate exposure
note we have `Female1` and `Female2`, these are already year beginning and year end estimates, already side-by side, so this is good. 
```{r}
E <-
  P |> 
  clean_names() |> 
  select(year, age, female1, female2) |> 
  mutate(exposure = (female1 + female2) / 2) |> 
  select(!starts_with("fem"))
```
## merge
```{r}
Fx <- 
  B |> 
  clean_names() |> 
  select(year, age, births = total) |> 
  left_join(E, by = join_by(year, age)) |> 
  mutate(Fx = births / exposure)
```

# 2. Calculate the General Fertility Rate, GFR
$$ GFR = 1000 \cdot \frac{\sum_{15}^{49}B_x}{\sum_{15}^{49}E_x} $$
```{r}
gfr <- 
  Fx |> 
  filter(between(age,15,49)) |> 
  #filter(age >= 15 & age <= 49)
  group_by(year) |> 
  summarize(GFR = 1000 * sum(births)/sum(exposure)) 
```
Curious about the trend?
```{r}
gfr |> 
  ggplot(aes(x = year,y = GFR)) +
  geom_line()
```

Why not 1975? The deviation in TFR is indeed small; this is coming from the HMD documentation, where we see that there was a universe definition change in that year, which stifles exposure calculations. The HMD has a slick way to account for this, but we don't have the info handy that we'd need.

```{r}
TFR <- 
  Fx |> 
  group_by(year) |> 
  summarize(TFR = sum(Fx)) 
TFR |> 
  ggplot(aes(x = year, y = TFR)) +
  geom_line() 
```

# 3. Make a plot of age-specific fertility rates

examples
```{r}
Fx |> 
  #filter(year %in% c(2000,2010, 2021)) |>
  filter(year > 1990) |> 
  ggplot(aes(x= age, y = Fx, color = year, group = year)) +
  geom_line(alpha = .6)
```
# 4. Calculate TFR
I'll just pipe it straight to a plot
```{r}
Fx |> 
  group_by(year) |> 
  summarize(TFR = sum(Fx),
            B = sum(births)) |> 
  ggplot(aes(x = year, y = TFR)) +
  geom_line() +
  geom_vline(xintercept = c(1936,1939))
```

5. calculate  MAB

$$ MAB = \frac{\sum F_x \cdot (x + .5)}{\sum F_x}$$


```{r}
mab <-
  Fx |> 
  group_by(year) |> 
  summarize(MABfx = sum(Fx * (age + .5)) / sum(Fx),
            MABb = sum(births * (age + .5)) / sum(births))
```
and plot code
```{r}
mab |> 
  pivot_longer(c(MABfx,MABb),names_to = "variant", values_to = "MAB") |> 
  ggplot(aes(x = year, y = MAB, color = variant)) +
  geom_line() +
  theme_minimal() +
  geom_vline(xintercept = c(1936,1939))
```

Curious how MAB and TFR move together?

```{r}
mab |> 
  left_join(TFR,by=join_by(year)) |> 
  ggplot(aes(x = TFR, y = MABfx, color = year)) +
  geom_point() +
  theme_minimal()
```
