==> 02_session.Rmd <==
---
title: "Session 2 Notes"
author: "Tim Riffe"
date: "2025-07-08"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# get the data:

We got these by right-clicking the "raw" button on github and copying the urls from the menu (not from the browser)
```{r}
library(tidyverse)
library(janitor)
D <- read_csv("https://github.com/timriffe/BSSD2025Module2/raw/refs/heads/master/data/ES_mort.csv.gz",show_col_types = FALSE) |> 
  clean_names()
```

# Calculate probabilities

So, we can use a value like .5 for $a_x$ in most ages, but not age 0 or in the open age group.

```{r}
mx_to_ax <- function(mx, sex, age){
  ax <-
    case_when(
    age == 0 & sex == "male" & mx < 0.02300 ~ 0.14929 - 1.99545 * mx,
    age == 0 & sex == "male" & mx >= 0.02300 & mx < 0.08307 ~ 0.02832 + 3.26021 * mx,
    age == 0 & sex == "male" & mx >= 0.08307 ~ 0.29915,
    age == 0 & sex == "female" & mx < 0.01724 ~ 0.14903 - 2.05527 * mx,
    age == 0 & sex == "female" & mx >= 0.01724 & mx < 0.06891 ~ 0.04667 + 3.88089 * mx,
    age == 0 & sex == "female" & mx >= 0.06891 ~ 0.31411,
    age == max(age) ~ 1 / mx,
    TRUE ~ .5)
  return(ax)
}
```
$$ q_x = \frac{m_x}{1 +(1 - a_x) \cdot m_x}$$
Now let's use it to calculate qx
```{r}
calc_qx <- function(mx, ax, age){
  qx <- mx / (1 + (1 - ax) * mx)
  qx <- if_else(age == max(age), 1, qx)
  return(qx)
}
```

# chain probabilities to make survivorship!

$$ \ell_x = radix \cdot \prod _0^{x-1} (1 - q_x)$$

```{r}
calc_lx <- function(qx, radix = 1e5){
  n  <- length(qx)
  lx <- c(1, cumprod(1 - qx))
  lx <- lx[1:n]
  lx <- radix * lx
  return(lx)
}
```


```{r}
D |> 
  group_by(year, sex) |> 
  fill(mx, .direction = "down") |> 
  mutate(mx = if_else(mx == 0,  .4, mx),
         ax = mx_to_ax(mx = mx, 
                       sex = sex, 
                       age = age),
         qx = calc_qx(mx = mx, 
                      ax = ax, 
                      age = age),
         lx = calc_lx(qx = qx, radix = 1e5),
         dx = lx * qx) |> 
  summarize(e0 = sum(dx * (age + ax)) / sum(dx)) |> 
  ggplot(aes(x = year, y = e0, color = sex)) +
  geom_line()
```


# exercises
Complete the lifetable by implementing these formulas:

Lx is what we all "lifetable exposure", person years lived in each age interval. If $a_x$ is .5 then this amounts to the trapezoid rule for discrete integration. But in ages 0 and 110 it isn't .5!
$$ L_x = \ell_x - (1 - a_x) \cdot d_x $$

Be sure that the last value of Lx is right:
$$ L_{110} = \ell_{110} \cdot a_{110}$$

Then an intermediate column:

$$ T_x = \sum_x^{110} L_x$$
This one is tricky to program. It's all the lifetable person-years lived beyond age x.

And finally remaining life expectancy, which has a value in each age, not just age 0:

$$ e_x = \frac{T_x}{\ell_x}$$

Now calculate the full lifetables for all subsets :-)
