==> 04_session.Rmd <==
---
title: "session 4 notes"
author: "Tim Riffe"
date: "2025-07-10"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Get data

```{r, message = FALSE}
library(tidyverse)
library(readr)
library(janitor)
source("https://raw.githubusercontent.com/timriffe/BSSD2025Module2/master/02_lifetables.R")
B <- read_csv("https://raw.githubusercontent.com/timriffe/BSSD2025Module2/master/data/ES_B2014.csv", show_col_types = FALSE) |> 
  mutate(sex = "female") |> 
  select(age, sex, births = total) 

D<- read_csv("https://raw.githubusercontent.com/timriffe/BSSD2025Module2/master/data/ES_D2014.csv", show_col_types = FALSE) |> 
     filter(year == 2014) |> 
  select(-open_interval) |> 
  pivot_longer(female:total, 
               names_to = "sex", 
               values_to = "deaths")

P <- read_csv("https://raw.githubusercontent.com/timriffe/BSSD2025Module2/master/data/ES_P.csv.gz", show_col_types = FALSE) |> 
  clean_names() |> 
  select(-open_interval) |> 
  filter(year== 2014) |> 
  pivot_longer(female1:total2, 
               names_to = "sex", 
               values_to = "pop") |> 
  mutate(period = parse_number(sex),
         sex = gsub('[[:digit:]]+', '', sex)) |> 
  pivot_wider(names_from = period, values_from = pop, names_prefix="pop")

ES2014 <- 
  left_join(P, D, by = join_by(year,sex, age)) |> 
  left_join(B, by = join_by(sex, age))

LT <- 
  ES2014 |> 
  mutate(exposure = (pop1 + pop2) / 2,
         mx = deaths / exposure,
         mx = if_else(mx == 0,1,mx)) |> 
  select(sex,age,mx) |> 
  group_by(sex) |> 
  group_modify(~lt_full(data = .x, radix = 1, group = .y)) |> 
  ungroup()
```

Calculate all the crude indicators for this short period

Note when we make B pick out total and not female to run this block.
```{r}
ES2014 |> 
  filter(sex == "total") |> 
  mutate(exposure = (pop1 + pop2) / 2) |> 
  summarize(E = sum(exposure),
            D = sum(deaths),
            B = sum(births, na.rm = TRUE),
            G = sum(pop2) - sum(pop1),
            P1 = sum(pop1),
            P2 = sum(pop2)) |> 
  mutate(CGR = G / E,
         CBR = B / E,
         CDR = D / E,
         CNMR = (P2 - (P1 + B - D)) / E,
         CNGR = CBR - CDR)
```

Let's build intuition on how growth rates can shape population age structure, using stable populations:

In general growth relates the present to the future:
$$ N(T) = N(0) \cdot e^{rT}$$ 
In reverse:
$$ N(-T) = N(0) \cdot e^{-Tr}$$

```{r}
r_values <- tibble(r = seq(-.03,.03,by=.005))
library(colorspace)
LT |> 
  filter(sex == "female") |> 
  select(age, Lx) |> 
  cross_join(r_values) |> 
  arrange(r, age) |> 
  group_by(r) |> 
  mutate(Cx = Lx * exp(-(age+.5) * r),
         Cx = Cx / sum(Cx)) |> 
  ggplot(aes(x = age, y = Cx, color = r, group = r)) +
  geom_line() +
  scale_color_binned_sequential() +
  theme_minimal()
```

# Let's gather what we need in order to do a population projection

1. Calculate ASFR (births / exposure)
```{r}
asfr <-
  ES2014 |> 
  filter(sex == "female") |> 
  select(-sex) |> 
  mutate(exposure = (pop1 + pop2) / 2) |> 
  arrange(age) |> 
  mutate(births = if_else(is.na(births),0,births),
         asfr = births / exposure) 
asfr |> 
ggplot(aes(x = age, y = asfr)) +
  geom_line()
```

2. Calculate Survival probabilities

$$ S_x = \frac{L_{x+1}}{L_x}$$
```{r}
Sx <-
  LT |> 
  filter(sex == "female") |> 
  select(age, Lx) |> 
  mutate(Sx = lead(Lx) / Lx) |> # changed mind here; removed 0 padding
  filter(!is.na(Sx)) # NEW
```

Turn these into the matrix diag. First step (next we make it become the subdiagonal)

```{r}
L <-
  Sx |> 
  select(-Lx) |> 
  mutate(age_to = age + 1) |> 
  pivot_wider(names_from = age,
              values_from = Sx,
              values_fill = 0) |> 
  column_to_rownames("age_to") |> 
  as.matrix()
```

pad the matrix

```{r}
Les <-
  0 |> 
  rbind(L) |> 
  cbind(0)
# cbind(rbind(0, L), 0)
rownames(Les) <- colnames(Les)
```

optional closeout

$$ \frac{T_\omega}{T_{\omega - 1}}$$ 
```{r, eval = FALSE}
Tx_closeout <-
  LT |> 
  filter(sex == "female",
  age >= 109) |> 
  pull(Tx)

Les[nrow(Les), ncol(Les)] <- Tx_closeout[2] / Tx_closeout[1]
```

# Fertility

Now, we have to account for fertility somehow, noting:
1. age groups need to be blended
2. we need to discount for the survival (in the age group) of mothers
3. we need to discount for girl births only
4. we need to discount for the survival of infants to age 0 on Dec 31.

Where does this come from?
```{r}
Fx  <- asfr$asfr
# PF  <- # anyone have a guess?
.4886 # Ken Wachter's favorite value for PF
SRB <- 1.05 # common perception
PF  <- 1 / (1 + SRB)
```
PF turns out to be quite important for this family of model, remember that

```{r}
L0 <- LT |> 
  filter(sex == "female",
         age == 0) |> 
  pull(Lx)

sx <-
  Sx |> 
  pull(Sx)
sx <- c(sx,0)

# induce narrow fertility:
TFR <- sum(Fx)

top_row <- L0 * PF * (Fx + lead(Fx,default = 0)) / 2 * sx


# second asfr schedule is meant to be narrower, to make a point
# about wave dissipation
Fx2 <- TFR * (Fx^3) / sum(Fx^3)
Fx2 <- c(Fx2[-c(1:5)],rep(0,5))
Fx3 <- c(rep(0,10),Fx2[1:101])
top_row2 <- L0 * PF * (Fx2 + lead(Fx,default = 0)) / 2 * sx
top_row3 <- L0 * PF * (Fx3 + lead(Fx,default = 0)) / 2 * sx
```

# Now impute this row into the Leslie matrix

First index is rows, second index is columns
```{r}
Les[1, ] <- top_row

# second copy with the narrower fertility distribution
Les2 <- Les
Les2[1, ] <- top_row2

Les3 <- Les
Les3[1, ] <- top_row3
```

How to project 1 year using this matrix
```{r}
N <- 
  P |> 
  filter(sex == "female") |> 
  pull(pop2)

plot(0:110, N, type = 'l')
lines(0:110, c( Les%*% N), col = "blue")
lines(0:110, c( Les%*% (Les%*% N)), col = "red")
```

# Let's do a longer-term population projection

To carry the population over longer time horizons, you can take matrix powers. See below: we can keep multiplying the Leslie matrix into the output of the previous multiplication h times, or we could just raise the Leslie matrix to the hth power. Using a matrix exponential. 
```{r, message = FALSE, echo = FALSE}
Les %*% (Les %*% (Les %*% N))

# install.packages("expm")
```

```{r}
plot(0:110, Fx3, type = 'l')
lines(0:110, Fx2, col = "blue")
lines(0:110, Fx, col = "red")
```



To do a longer-run population projection _under constant conditions_, you can set it up like so:
```{r}
library(expm)
horizon <- tibble(h = 0:200)

proj <-
  P |> 
  filter(sex == "female") |> 
  select(year, age, p0 = pop2) |> 
  # this line induces waviness in the starting population,
  # for the experiment about wave dissipation
  #mutate(p02 = p0 * (sin(age /10) + 1.3)) |> 
  cross_join(horizon) |> 
  arrange(h, age) |> 
  group_by(h) |> 
  mutate(pop = c((Les %^% h[1]) %*% p0 ),
         pop2 = c((Les2 %^% h[1]) %*% p0 ),
         pop3 = c((Les3 %^% h[1]) %*% p0 ))
```

# Exercises

1. Plot the path of CGR out to the end of the projection. What do you see?

```{r}
cgr <-
  proj |> 
  group_by(h) |> 
  summarize(P1 = sum(pop),
            P2 = sum(pop2),
            P3 = sum(pop3)) |> 
  mutate(E1 = (P1 + lead(P1)) / 2,
         E2 = (P2 + lead(P2)) / 2,
         E3 = (P3 + lead(P3)) / 2,
         G1 = c(diff(P1),NA),
         G2 = c(diff(P2),NA),
         G3 = c(diff(P3),NA)) |> 
  filter(!is.na(E1)) |> 
  mutate(CGR1 = G1 / E1,
         CGR2 = G2 / E2,
         CGR3 = G3 / E3)
```

Tidy redux of the above
```{r}
cgr <-
  proj |> 
  pivot_longer(pop:pop3, names_to = "variant", values_to = "pop") |> 
  group_by(variant, h) |> 
  summarize(P = sum(pop), .groups = "drop") |> 
  group_by(variant) |> 
  mutate(E = (P + lead(P)) / 2,
         G = c(diff(P),NA)) |> 
  ungroup() |> 
  filter(!is.na(E)) |> 
  mutate(CGR = G / E)
```


```{r}
cgr |> 
  select(variant, h, CGR) |> 
  ggplot(aes(x = h, y = CGR, color = variant)) +
  geom_line() +
  theme_minimal()
```

2. Make a plot that shows the (scaled) population structure every 20 years to get a sense of how long it takes to stabilize. Which fertility distribution (1 or variant 2?) seems to have a rougher path?

```{r}
proj |> 
  pivot_longer(pop:pop3, names_to = "variant", values_to = "pop") |> 
  filter(h %% 20 == 0) |> 
  group_by(h, variant) |> 
  select(h,age,variant,pop) |> 
  group_by(h,variant) |> 
  mutate(cx = 100 * pop / sum(pop)) |> 
  ggplot(aes(x=age,y=cx,color = variant)) +
  geom_line() +
  facet_wrap(~h) +
  theme_minimal()
```


# how to get the stable parameters from a Leslie matrix

```{r}
r <- eigen(Les)$values[1] |> Re() |> log()
```

```{r}
Sx |> 
  select(age, Lx) |> 
  mutate(Cx = Lx * exp(-(age+.5) * r),
         Cx = 100 * Cx / sum(Cx)) |> 
  ggplot(aes(x = age, y = Cx)) +
  geom_line()
```
